requirements.txt
streamlit==1.39.0
requests==2.32.3
pandas==2.2.3
plotly==5.24.1
python-dotenv==1.0.1
numpy==2.1.2
